Fix small bugs around the directory/file install.
For the directory install, if the path is relative and the
extension expects to run configure, the path to configure
is bogus as the subprocess runs the extension directory.
For the file install, if the extension expects to run
configure before make, Makefile does not necessarily exit,
so discerning the extension's directory by any of 'Makefile',
'makefile' or 'GNUMakefile' was not enough.